% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb-cores.R
\name{assert_cores}
\alias{assert_cores}
\title{Check number of cores}
\usage{
assert_cores(ncores)
}
\arguments{
\item{ncores}{Number of cores to check. Make sure is not larger than
\code{getOption("bigstatsr.ncores.max")} (number of logical cores by default).}
}
\description{
Check number of cores
}
\examples{
assert_cores(2)

}
