% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigparallelr-package.R
\docType{package}
\name{bigparallelr-package}
\alias{bigparallelr}
\alias{bigparallelr-package}
\title{bigparallelr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
