% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb-cores.R
\name{get_blas_ncores}
\alias{get_blas_ncores}
\alias{set_blas_ncores}
\title{Number of cores used by BLAS (matrix computations)}
\usage{
get_blas_ncores()

set_blas_ncores(ncores)
}
\arguments{
\item{ncores}{Number of cores to set for BLAS.}
}
\description{
Number of cores used by BLAS (matrix computations)
}
\examples{
get_blas_ncores()
}
