% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split-parapply.R
\name{plus}
\alias{plus}
\title{Add}
\usage{
plus(...)
}
\arguments{
\item{...}{Multiple arguments to be added together.}
}
\value{
\code{Reduce('+', list(...))}
}
\description{
Wrapper around \code{Reduce} to add multiple arguments. Useful
}
\examples{
plus(1:3, 4:6, 1:3)
}
