% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-parallel.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{makeCluster}
\alias{stopCluster}
\alias{registerDoParallel}
\alias{lock}
\alias{unlock}
\alias{blas_set_num_threads}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{RhpcBLASctl}{\code{\link[RhpcBLASctl]{blas_set_num_threads}}}

  \item{doParallel}{\code{\link[doParallel]{registerDoParallel}}}

  \item{flock}{\code{\link[flock]{lock}}, \code{\link[flock]{unlock}}}

  \item{parallel}{\code{\link[parallel]{makeCluster}}, \code{\link[parallel]{stopCluster}}}
}}

