% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{rows_along}
\alias{rows_along}
\alias{cols_along}
\alias{seq_range}
\title{Sequence generation}
\usage{
rows_along(x)

cols_along(x)

seq_range(lims)
}
\arguments{
\item{x}{Any object on which you can call \code{nrow()} and \code{ncol()}.}

\item{lims}{Vector of size 2 (or more, but only first 2 values will be used).}
}
\description{
\itemize{
\item \code{rows_along(x)}: \code{seq_len(nrow(x))}
\item \code{cols_along(x)}: \code{seq_len(ncol(x))}
\item \code{seq_range(lims)}: \code{seq(lims[1], lims[2])}
}
}
\examples{
X <- matrix(1:6, 2, 3)
dim(X)
rows_along(X)
cols_along(X)

seq_range(c(3, 10))

}
\keyword{internal}
