% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{split_len}
\alias{split_len}
\title{Split length in blocks}
\usage{
split_len(total_len, block_len, nb_split = ceiling(total_len/block_len))
}
\arguments{
\item{total_len}{Length to split.}

\item{block_len}{Maximum length of each block.}

\item{nb_split}{Number of blocks. Default uses the other 2 parameters.}
}
\value{
A matrix with 3 columns \code{lower}, \code{upper} and \code{size}.
}
\description{
Split length in blocks
}
\examples{
split_len(10, block_len = 3)
split_len(10, nb_split = 3)

}
