% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigparallelr-package.R
\docType{package}
\name{bigparallelr-package}
\alias{bigparallelr}
\alias{bigparallelr-package}
\title{bigparallelr: Easy Parallel Tools}
\description{
Utility functions for easy parallelism in R. Include some reexports
    from other packages, utility functions for splitting and parallelizing over
    blocks, and choosing and setting the number of cores used.
}
\author{
\strong{Maintainer}: Florian Privé \email{florian.prive.21@gmail.com}

}
\keyword{internal}
