% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigparallelr-package.R
\docType{package}
\name{bigparallelr-package}
\alias{bigparallelr}
\alias{bigparallelr-package}
\title{bigparallelr: Parallel Tools}
\description{
Utility functions for easy parallelism in R.
}
\author{
\strong{Maintainer}: Florian Privé \email{florian.prive.21@gmail.com}

}
\keyword{internal}
